Refactor object types
