+++ src/argument_retriever.hpp
@@ -13,8 +13,8 @@
 #include <vector>
 
 #include "exclusion/common.hpp"
+#include "object.hpp"
 #include "object_store.hpp"
-#include "object_view.hpp"
 #include "traits.hpp"
 
 namespace ddwaf {
