+++ src/condition/cmdi_detector.cpp
@@ -25,8 +25,8 @@
 #include "exclusion/common.hpp"
 #include "iterator.hpp"
 #include "log.hpp"
+#include "object.hpp"
 #include "object_type.hpp"
-#include "object_view.hpp"
 #include "platform.hpp"
 #include "tokenizer/shell.hpp"
 #include "transformer/common/cow_string.hpp"
