+++ src/condition/exists.cpp
@@ -16,8 +16,8 @@
 #include "condition/exists.hpp"
 #include "exception.hpp"
 #include "exclusion/common.hpp"
+#include "object.hpp"
 #include "object_type.hpp"
-#include "object_view.hpp"
 #include "utils.hpp"
 
 namespace ddwaf {
