+++ src/condition/shi_detector.cpp
@@ -17,8 +17,8 @@
 #include "condition/structured_condition.hpp"
 #include "exclusion/common.hpp"
 #include "log.hpp"
+#include "object.hpp"
 #include "object_type.hpp"
-#include "object_view.hpp"
 #include "tokenizer/shell.hpp"
 #include "utils.hpp"
