+++ src/iterator.cpp
@@ -15,7 +15,6 @@
 #include "iterator.hpp"
 #include "object.hpp"
 #include "object_type.hpp"
-#include "object_view.hpp"
 #include "utils.hpp"
 
 namespace ddwaf {
