+++ src/processor/extract_schema.cpp
@@ -23,7 +23,6 @@
 #include "object.hpp"
 #include "object_store.hpp"
 #include "object_type.hpp"
-#include "object_view.hpp"
 #include "processor/base.hpp"
 #include "processor/extract_schema.hpp"
 #include "scanner.hpp"
