+++ src/transformer/manager.cpp
@@ -7,7 +7,6 @@
 #include <string_view>
 
 #include "object.hpp"
-#include "object_view.hpp"
 #include "transformer/base.hpp"
 #include "transformer/base64_decode.hpp"
 #include "transformer/base64_encode.hpp"
